--- SecretSanta/SecretSanta/SecretSanta.sln
@@ -1,74 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29521.150
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{71196D9A-D1FA-469B-86BD-08620BD5BC85}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Data", "src\SecretSanta.Data\SecretSanta.Data.csproj", "{72F070FE-B8A5-4A77-B8ED-764068CAEFD3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Business", "src\SecretSanta.Business\SecretSanta.Business.csproj", "{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Api", "src\SecretSanta.Api\SecretSanta.Api.csproj", "{C33A221C-CDAB-4245-A9A4-B32A5AB9A690}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Web", "src\SecretSanta.Web\SecretSanta.Web.csproj", "{E5663FEA-B7B2-4C84-B6F7-44744A427C87}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Api.Tests", "tests\SecretSanta.Api.Tests\SecretSanta.Api.Tests.csproj", "{74C518FE-F22E-4E95-9198-0C0B73470BBF}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{964F8A4C-44A5-4640-A3CC-7BB329E66995}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Business.Tests", "tests\SecretSanta.Business.Tests\SecretSanta.Business.Tests.csproj", "{720F3E98-DA49-4407-B41F-13BA3D81D88C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Data.Tests", "tests\SecretSanta.Data.Tests\SecretSanta.Data.Tests.csproj", "{493F624D-A13A-46C4-A5CC-A861BD628691}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{72F070FE-B8A5-4A77-B8ED-764068CAEFD3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{72F070FE-B8A5-4A77-B8ED-764068CAEFD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{72F070FE-B8A5-4A77-B8ED-764068CAEFD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{72F070FE-B8A5-4A77-B8ED-764068CAEFD3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C33A221C-CDAB-4245-A9A4-B32A5AB9A690}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C33A221C-CDAB-4245-A9A4-B32A5AB9A690}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C33A221C-CDAB-4245-A9A4-B32A5AB9A690}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C33A221C-CDAB-4245-A9A4-B32A5AB9A690}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Release|Any CPU.Build.0 = Release|Any CPU
-		{74C518FE-F22E-4E95-9198-0C0B73470BBF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{74C518FE-F22E-4E95-9198-0C0B73470BBF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{74C518FE-F22E-4E95-9198-0C0B73470BBF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{74C518FE-F22E-4E95-9198-0C0B73470BBF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{720F3E98-DA49-4407-B41F-13BA3D81D88C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{720F3E98-DA49-4407-B41F-13BA3D81D88C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{720F3E98-DA49-4407-B41F-13BA3D81D88C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{720F3E98-DA49-4407-B41F-13BA3D81D88C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{493F624D-A13A-46C4-A5CC-A861BD628691}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{493F624D-A13A-46C4-A5CC-A861BD628691}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{493F624D-A13A-46C4-A5CC-A861BD628691}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{493F624D-A13A-46C4-A5CC-A861BD628691}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{72F070FE-B8A5-4A77-B8ED-764068CAEFD3} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
-		{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
-		{C33A221C-CDAB-4245-A9A4-B32A5AB9A690} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
-		{E5663FEA-B7B2-4C84-B6F7-44744A427C87} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
-		{74C518FE-F22E-4E95-9198-0C0B73470BBF} = {964F8A4C-44A5-4640-A3CC-7BB329E66995}
-		{720F3E98-DA49-4407-B41F-13BA3D81D88C} = {964F8A4C-44A5-4640-A3CC-7BB329E66995}
-		{493F624D-A13A-46C4-A5CC-A861BD628691} = {964F8A4C-44A5-4640-A3CC-7BB329E66995}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E2234040-21B9-4AF6-9A6A-3ABE8D5F86E0}
-	EndGlobalSection
-EndGlobal
