--- SecretSanta/SecretSanta/src/SecretSanta.Business/gift.cs
@@ -1,24 +1,0 @@
-using System;
-using System.Collections.Generic;
-
-
-namespace SecretSanta.Business
-{
-    public class Gift
-    {
-        public int Id {get;}
-        public string Title {get; set;}
-        public string Description {get; set;}
-        public string Url {get; set;}
-        public User User {get; set;}
-
-        public Gift(int id, string title, string description, string url, User user)
-        {
-            Id = id;
-            Title = title ?? throw new ArgumentNullException(nameof(title));
-            Description = description ?? throw new ArgumentNullException(nameof(description));
-            Url = url ?? throw new ArgumentNullException(nameof(url));
-            User = user ?? throw new ArgumentNullException(nameof(user));
-        }
-    }
-}
