+++ SecretSanta/src/SecretSanta.Business/GlobalSuppressions.cs
